UI shows all cities

+++ Assets/Scripts/Building.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 public class Building : ScriptableObject {
   public new string name;
   public Vector2Int size;
-  public int buildingHeight;
-  [HideInInspector]public int height;
+  public float buildingHeight;
+  [HideInInspector]public float height;
 
   public Material material;
